refactor: add the alert component on context
