+++ utils/share.ts
@@ -1,0 +1,8 @@
+import { APP_LINK } from "@/constants/links"
+import { Share } from "react-native"
+
+export const onShare = (username: string = "me") => {
+  Share.share({
+    message: `Hey! Join ${username} using Wavv and keep up to date with things they do! Download the app here: ${APP_LINK}`,
+  })
+}
